Restructure test files
This moves to using the compile-tests structure I had originally
intended, where we include some set of test font files, and for each of
those files we include a set of expected successes and expected
failures.

I'd put this off, but I'm finding it hard to add simple tests of
compilation success, and this is a good way of doing that.